Improve WorkMapTable stories
