+++ turboui/src/WorkMap/tests/mockData.ts
@@ -1,7 +1,7 @@
 import WorkMap from "../components";
 import { TimeframeSelector } from "../../TimeframeSelector";
 import { PrivacyIndicator } from "../../PrivacyIndicator";
-import { currentQuarter, currentYear } from "../../utils/timeframes";
+import { currentMonth, currentQuarter, currentYear } from "../../utils/timeframes";
 
 function genAvatar(id: string) {
   return `https://images.unsplash.com/${id}?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80`;
@@ -470,3 +470,269 @@ export const mockItems: WorkMap.Item[] = [
   },
 ];
 
+export const onlyGoals: WorkMap.Item[] = [
+  {
+    id: "goal-g1",
+    parentId: null,
+    type: "goal",
+    name: "Increase NPS score",
+    status: "caution",
+    progress: 60,
+    space: { id: "space-cs", name: "Customer Success" },
+    spacePath: "#",
+    owner: people.alex,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: null,
+    closedAt: null,
+    nextStep: "Send Q3 survey",
+    privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+    timeframe: getTimeframe(currentQuarter()),
+    children: [
+      {
+        id: "goal-g1-1",
+        parentId: "goal-g1",
+        type: "goal",
+        name: "Launch new CSAT tool",
+        status: "achieved",
+        progress: 80,
+        space: { id: "space-cs", name: "Customer Success" },
+        spacePath: "#",
+        owner: people.sophia,
+        ownerPath: "#",
+        itemPath: "#",
+        isNew: false,
+        completedOn: "2025-08-15T00:00:00.000Z",
+        closedAt: null,
+        nextStep: "Enable feedback widget",
+        privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+        timeframe: getTimeframe(currentMonth()),
+        children: [],
+      },
+      {
+        id: "goal-g1-2",
+        parentId: "goal-g1",
+        type: "goal",
+        name: "Reduce churn by 10%",
+        status: "missed",
+        progress: 50,
+        space: { id: "space-cs", name: "Customer Success" },
+        spacePath: "#",
+        owner: people.jennifer,
+        ownerPath: "#",
+        itemPath: "#",
+        isNew: false,
+        completedOn: null,
+        closedAt: null,
+        nextStep: "Analyze churn data",
+        privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+        timeframe: getTimeframe(currentQuarter()),
+        children: [],
+      },
+    ],
+  },
+  {
+    id: "goal-g2",
+    parentId: null,
+    type: "goal",
+    name: "Expand to APAC region",
+    status: "partial",
+    progress: 30,
+    space: { id: "space-growth", name: "Growth" },
+    spacePath: "#",
+    owner: people.sophia,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: null,
+    closedAt: null,
+    nextStep: "Research local partners",
+    privacy: "secret" as PrivacyIndicator.PrivacyLevels,
+    timeframe: getTimeframe(currentYear()),
+    children: [
+      {
+        id: "goal-g2-1",
+        parentId: "goal-g2",
+        type: "goal",
+        name: "Localize product",
+        status: "pending",
+        progress: 40,
+        space: { id: "space-growth", name: "Growth" },
+        spacePath: "#",
+        owner: people.alex,
+        ownerPath: "#",
+        itemPath: "#",
+        isNew: false,
+        completedOn: null,
+        closedAt: null,
+        nextStep: "Hire translators",
+        privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+        timeframe: getTimeframe(currentQuarter()),
+        children: [],
+      },
+      {
+        id: "goal-g2-2",
+        parentId: "goal-g2",
+        type: "goal",
+        name: "APAC marketing campaign",
+        status: "dropped",
+        progress: 25,
+        space: { id: "space-growth", name: "Growth" },
+        spacePath: "#",
+        owner: people.jennifer,
+        ownerPath: "#",
+        itemPath: "#",
+        isNew: false,
+        completedOn: null,
+        closedAt: null,
+        nextStep: "Draft campaign assets",
+        privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+        timeframe: getTimeframe(currentQuarter()),
+        children: [],
+      },
+    ],
+  },
+];
+
+export const onlyProjects: WorkMap.Item[] = [
+  {
+    id: "project-1",
+    parentId: null,
+    type: "project",
+    name: "Mobile app v2",
+    status: "dropped",
+    progress: 70,
+    space: { id: "space-mobile", name: "Mobile" },
+    spacePath: "#",
+    owner: people.sophia,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: null,
+    closedAt: null,
+    nextStep: "Start beta testing",
+    privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+    timeframe: {
+      startDate: "2025-01-10T00:00:00.000Z",
+      endDate: "2025-09-30T00:00:00.000Z",
+      type: "days",
+    },
+    children: [],
+  },
+  {
+    id: "project-2",
+    parentId: null,
+    type: "project",
+    name: "API refactor",
+    status: "paused",
+    progress: 40,
+    space: { id: "space-backend", name: "Backend" },
+    spacePath: "#",
+    owner: people.alex,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: null,
+    closedAt: null,
+    nextStep: "Migrate endpoints",
+    privacy: "confidential" as PrivacyIndicator.PrivacyLevels,
+    timeframe: {
+      startDate: "2025-02-15T00:00:00.000Z",
+      endDate: "2025-10-31T00:00:00.000Z",
+      type: "days",
+    },
+    children: [],
+  },
+  {
+    id: "project-3",
+    parentId: null,
+    type: "project",
+    name: "Website redesign",
+    status: "achieved",
+    progress: 50,
+    space: { id: "space-design", name: "Design" },
+    spacePath: "#",
+    owner: people.jennifer,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: "2025-07-31T00:00:00.000Z",
+    closedAt: null,
+    nextStep: "Finalize homepage",
+    privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+    timeframe: {
+      startDate: "2025-03-01T00:00:00.000Z",
+      endDate: "2025-07-31T00:00:00.000Z",
+      type: "days",
+    },
+    children: [],
+  },
+];
+
+export const onlyCompleted: WorkMap.Item[] = [
+  {
+    id: "goal-10",
+    parentId: null,
+    type: "goal",
+    name: "Achieve Series B funding",
+    status: "completed",
+    progress: 100,
+    space: { id: "space-finance", name: "Finance" },
+    spacePath: "#",
+    owner: people.alex,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: "2025-06-01T00:00:00.000Z",
+    closedAt: "2025-06-01T00:00:00.000Z",
+    nextStep: "",
+    privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+    timeframe: getTimeframe(currentYear()),
+    children: [],
+  },
+  {
+    id: "goal-11",
+    parentId: null,
+    type: "goal",
+    name: "Hit 100k MAU",
+    status: "completed",
+    progress: 100,
+    space: { id: "space-growth", name: "Growth" },
+    spacePath: "#",
+    isNew: false,
+    completedOn: "2025-05-15T00:00:00.000Z",
+    closedAt: "2025-05-01T00:00:00.000Z",
+    owner: people.sophia,
+    ownerPath: "#",
+    itemPath: "#",
+    nextStep: "",
+    privacy: "internal" as PrivacyIndicator.PrivacyLevels,
+    timeframe: getTimeframe(currentQuarter()),
+    children: [],
+  },
+  {
+    id: "project-10",
+    parentId: null,
+    type: "project",
+    name: "Migrate to Vite",
+    status: "completed",
+    progress: 100,
+    space: { id: "space-eng", name: "Engineering" },
+    spacePath: "#",
+    owner: people.jennifer,
+    ownerPath: "#",
+    itemPath: "#",
+    isNew: false,
+    completedOn: "2025-04-01T00:00:00.000Z",
+    closedAt: "2025-04-01T00:00:00.000Z",
+    nextStep: "",
+    privacy: "secret" as PrivacyIndicator.PrivacyLevels,
+    timeframe: {
+      startDate: "2025-01-15T00:00:00.000Z",
+      endDate: "2025-04-01T00:00:00.000Z",
+      type: "days",
+    },
+    children: [],
+  },
+];
